+++ src/main/java/org/mbari/mxm/provider_client/MxmProviderClient.java
@@ -14,11 +14,9 @@ public interface MxmProviderClient {
 
   ProviderApiType apiType();
 
-  PingResponse ping() throws ProviderPingException;
+  PingResponse ping(MxmInfo pl) throws ProviderPingException;
 
   GeneralInfoResponse getGeneralInfo();
-
-  MxmInfo postMxmInfo(MxmInfo pl);
 
   AssetClassesResponse getAssetClasses();